requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
dj-email-url==1.0.6
Django==5.2
django-cache-url==3.4.5
django-crispy-forms==2.4
django-phonenumber-field==8.3.0
environs==14.3.0
idna==3.11
marshmallow==4.0.1
phonenumbers==9.0.18
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
